Added mean_absolute_error function for changing the framwork

+++ ivy/functional/frontends/tensorflow/metrics.py
@@ -147,7 +147,7 @@ def kl_divergence(y_true, y_pred):
 
 
 def mean_absolute_error(y_true, y_pred):
-    return ivy.mean(ivy.abs(y_true - y_pred), axis=-1)
+    return ivy.mean(ivy.abs(y_true - y_pred), axis=-1).astype(y_true.dtype)
 
 
 mae = mean_absolute_error
